+++ lib/lintConfig.ts
@@ -162,6 +162,7 @@ const defaultLintRuleConfig: Required<LintRuleConfig<FullLintConfigValue>> = {
 			// dynamicpagelist: 2,
 			// inputbox: 2,
 			// references: 2,
+			// seo: 2,
 		},
 	],
 	'obsolete-attr': [1],