+++ src/multiLine/index.ts
@@ -1,4 +1,5 @@
 import {Token} from '../index';
+import {gapped} from '../../mixin/gapped';
 import type {AttributesToken, ExtToken} from '../../internal';
 
 /* NOT FOR BROWSER */
@@ -13,6 +14,7 @@ import {clone} from '../../mixin/clone';
  *
  * 逐行解析的扩展标签
  */
+@gapped()
 export abstract class MultiLineToken extends Token {
 	declare readonly name: string;
 
@@ -40,11 +42,6 @@ export abstract class MultiLineToken extends Token {
 	/** @private */
 	override text(): string {
 		return super.text('\n').replace(/\n\s*\n/gu, '\n');
-	}
-
-	/** @private */
-	override getGaps(): number {
-		return 1;
 	}
 
 	/** @private */