+++ base.ts
@@ -143,6 +143,8 @@ export type TokenTypes = 'root'
 	| 'converter-rule-to'
 	| 'converter-rule-from'
 	| 'param-line'
+	| 'param-line-key'
+	| 'param-line-value'
 	| 'imagemap-link'
 
 	/* NOT FOR BROWSER */