+++ src/atom.ts
@@ -27,6 +27,8 @@ const atomTypes = [
 	'invoke-module',
 	'template-name',
 	'link-target',
+	'param-line-key',
+	'param-line-value',
 ] as const;
 
 declare type AtomTypes = typeof atomTypes[number];