+++ src/nested.ts
@@ -9,6 +9,7 @@ import {NoincludeToken} from './nowiki/noinclude';
 import type {
 	Config,
 	LintError,
+	TokenTypes,
 } from '../base';
 import type {CommentToken, AttributesToken, IncludeToken, ArgToken, TranscludeToken} from '../internal';
 
@@ -22,7 +23,7 @@ import {clone} from '../mixin/clone';
 
 declare type Child = ExtToken | NoincludeToken | CommentToken | IncludeToken | ArgToken | TranscludeToken;
 
-const childTypes = new Set(['comment', 'include', 'arg', 'template', 'magic-word']),
+const childTypes = new Set<TokenTypes>(['comment', 'include', 'arg', 'template', 'magic-word']),
 	lintRegex = /* #__PURE__ */ (() => [false, true].map(article => {
 		const noinclude = article ? 'includeonly' : 'noinclude';
 		// eslint-disable-next-line @typescript-eslint/no-unused-expressions