+++ src/multiLine/index.ts
@@ -1,3 +1,4 @@
+import Parser from '../../index';
 import {Token} from '../index';
 import {gapped} from '../../mixin/gapped';
 import type {AttributesToken, ExtToken} from '../../internal';
@@ -32,6 +33,20 @@ export abstract class MultiLineToken extends Token {
 
 	override get type(): 'ext-inner' {
 		return 'ext-inner';
+	}
+
+	/** @class */
+	constructor(
+		wikitext?: string,
+		config = Parser.getConfig(),
+		accum: Token[] = [],
+		acceptable?: WikiParserAcceptable,
+	) {
+		super(undefined, config, accum, acceptable);
+		if (wikitext) {
+			const newConfig = {...config, excludes: [...config.excludes, 'list']};
+			this.safeAppend(wikitext.split('\n').map(line => new Token(line, newConfig, accum)));
+		}
 	}
 
 	/** @private */