+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!-- markdownlint-disable first-line-h1 -->
 ## v1.45.0
 
-*2026-06-21*
+*2026-06-22*
 
 **Added**
 
@@ -15,6 +15,7 @@
 - [`TagToken#findMatchingTag`](https://github.com/bhsd-harry/wikiparser-node/wiki/TagToken-%28EN%29#findmatchingtag) after [`Token#buildLists`](https://github.com/bhsd-harry/wikiparser-node/wiki/Token-%28EN%29#buildlists) is executed
 - Missing `ImagemapLinkToken#cloneNode` method
 - JSON schema for `<templatedata>`
+- Parsing of `<poll>` extension tags
 
 ## v1.44.0